Callback function for when no results are found
Added an option to include a callback that is executed when not results
are found.  The option is
'no_results_callback(no_results_html_line,text)'.
'no_results_html_line' is the list item with the label saying 'no
results found' and the 'text' is the entered text that could not be
found.