Fixing bug related to choice destroy click.
This particular issue was tricky to debug, but essentially there was
an issue with clicking to destroy a choice in the multi select.

The current bug can be reproduced by following the steps here:
https://github.com/harvesthq/chosen/issues/346

I tried to solve it by forcing the field to blur after the choice
delete link was clicked.  However, there was a race condition
involving @pending_destroy_click, container's mouseup event, and the
destroy link's click event.