Search matches optgroup
This addition will allow to have search/filter match on 'optgroup' as
well. This way if the group label contains characters of search/filter,
show all 'options' in that 'optgroup'.

This is used as an option in the chosen object, by default it should
not search for optgroup. If you want to enable it, use the
allow_optgroup_search flag and set it to true. Example attached as well.