Updated the coffee script source with the new code, fixed bugs with the way the matching worked and added score.js source
Note that for some reason every time I tried to compile the score.js
from coffee script source format, which wrapped it in a function
literal, or if I made it a function of the chosen object, it would
cause the script to run for a long time. It only seems to work properly
when left as-is and included as-is. This can probably be fixed if
somebody wants to spend the time figuring out how to refactor it to
work that way.