Use the placeholder attribute as default text when it is availabe
I'm not sure why you chose to use @data-placeholder in the first place but there is nothing wrong with @placeholder. This attribute is now supported in most major browser.

I suggest to also check that attribute to find the default text of the select. My patch is backward compatible with @data-placeholder, of course.