Better support for hidden selects when applying Chosen
For jQuery, we first test the css('width') of a select. This always
returns a value for visible selects, but it will return false
for hidden selects with no style set (inline or class based).

For Prototype, we first check the getStyle('width') of a select. This
behaves much the same way as jQuery's css('width'), fortunately.