Timing of 'chosen:showing_dropdown' fixed
Currently the event is triggered before dropdown LI's
are created. This commit corrects it by moving the
trigger event at end of method. A JSFIDDLE was created
to demonstrate the issue.

http://jsfiddle.net/lagrz/3A5jq/

By moving the event it enables developers to add
whatever additional code or styling to the dropdown
menu options such as images.