Use style width if select list is hidden and width option not set.
If the select list or its container is hidden, width is set to 0px. I know that we can set the width option on the chosen() method (ie, chosen({width: '100px'}), but that forces us to use a chosen() statement for each and every hidden select vs being able to simply call $('.chosen-select').chosen() (and being able to control look and feel with page level css).

If there is no width option and offsetWidth is 0px, then chosen should try to use the width style attribute on the select list.