Fixed result-selected class being added to more than one option after the first use of the drop down.
The result-selected class can be sometimes used to hide the currently selected option to avoid redundancy in the options being shown. But because of the issue that this change fixes, in 3 consequtive uses, after the third use you will not be able to see the previous to selections in the drop down menu.