allow native select interfaces for iOS and android
Allow chosen to take advantage of the arguably superior native
selection interface on iOS and Android, while retaining the look and
feel of chosen for inactive states.

This is accomplished through a sort-of trick where a form element's
opacity can be set to 0 so that it is unseen while still remaining
interactive.

In this case, we can render the `.chosen-select` node in front of the
`.chosen-choices` or `.chosen-single` node, but with an opacity of 0,
so that it is invisible but interactive.

All of this behavior can be toggled with a `use_native_interface`
option, but is enabled by default for iOS and Android through
userAgent string matching.