fixing `.chosen-select` height bug on webkit
The opaque `.chosen-select` node was not exactly the same height as the
`.chosen-container` node behind it, which allowed the possibility of
some tap events on the edges to make it down to the other chosen nodes,
triggering the non-native behavior. This change should prevent that from
happening.

Webkit will not respect the `height` style for `select` nodes with the
default `-webkit-appearance: menulist` style. We can change this by
changing `-webkit-appearance` to `menulist-button`, which does respect
the `height` style.