Use content-box for text inputs
It's normally the default anyway but it's becoming very popular to set
* { box-sizing: content-box; } because it fixes so many annoying things.
Unfortunately, it can cause a problem when height and top or bottom
padding are set. In the case of this element, if border-box has been
set, we will only have 5px of vertical space left. So make sure it's
using content-box.