Allow detached head when parsing master ref.
If we have checked out a commit instead of a branch name, git is in what’s known as “detached head” state.

This is what Travis does all the time.

This means that we were never considered to be on the branch “master”, and therefore were never publishing bower-chosen.

We can tweak the git command we’re calling to support either checked out branch names or detached heads — as long as we’re at the commit for a commit that is the head of master, we’re peachykeen!