Add new trigger buildresultslist
This is simply a pass through to the winnow_results function. There is
an A11Y reason behind this, when a page loads, and chosen is activated,
the results list is empty, so the aria labels are pointing to an empty
list, which is flagged as an error in many web scanners. As soon as the
results list is built (the first time the field is activated) the
scanners stop complaining.

The reason I chose to add it as a trigger, is for dynamic select lists.
Normally, chosen:updated would be used when options are inserted, the
problem is that you can't use chosen:updated until the results list has
been built once, because it doesn't actually generate the results list.

I have decided NOT to force a build on initialization to avoid adding
overhead for those who don't need the A11Y work. Instead, the trigger
needs to be called in your script wherever Chosen is initialized.

If there is a better way to build the options list than this, please let
me know, I am admittedly new to the Chosen source code.