Super fast chosen
This code of chosen can handle selects with options like 50,000 and more easily. I've tested it with 50,000 data it's working like anything.
This code has been written on 0.9.11 version of chosen, i'm still learning coffeescript so i need to put these changes in head rev. The idea is not to create all li elements in chosen instead create them on demand, with scroll or search.